# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * High-Protein Vegetarian Macro Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProfileGoal = (typeof ProfileGoal)[keyof typeof ProfileGoal];

export const ProfileGoal = {
  muscle_gain: "muscle_gain",
  fat_loss: "fat_loss",
  maintenance: "maintenance",
} as const;

export type ProfileActivityLevel =
  (typeof ProfileActivityLevel)[keyof typeof ProfileActivityLevel];

export const ProfileActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
} as const;

export interface Profile {
  id: number;
  name: string;
  weightKg: number;
  /** @nullable */
  heightCm: number | null;
  /** @nullable */
  ageYears: number | null;
  goal: ProfileGoal;
  activityLevel: ProfileActivityLevel;
  dietaryRestrictions: string[];
  preferredProteins: string[];
  createdAt: string;
  updatedAt: string;
}

export type CreateProfileBodyGoal =
  (typeof CreateProfileBodyGoal)[keyof typeof CreateProfileBodyGoal];

export const CreateProfileBodyGoal = {
  muscle_gain: "muscle_gain",
  fat_loss: "fat_loss",
  maintenance: "maintenance",
} as const;

export type CreateProfileBodyActivityLevel =
  (typeof CreateProfileBodyActivityLevel)[keyof typeof CreateProfileBodyActivityLevel];

export const CreateProfileBodyActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
} as const;

export interface CreateProfileBody {
  name: string;
  weightKg: number;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  ageYears?: number | null;
  goal: CreateProfileBodyGoal;
  activityLevel: CreateProfileBodyActivityLevel;
  dietaryRestrictions?: string[];
  preferredProteins?: string[];
}

export type UpdateProfileBodyGoal =
  (typeof UpdateProfileBodyGoal)[keyof typeof UpdateProfileBodyGoal];

export const UpdateProfileBodyGoal = {
  muscle_gain: "muscle_gain",
  fat_loss: "fat_loss",
  maintenance: "maintenance",
} as const;

export type UpdateProfileBodyActivityLevel =
  (typeof UpdateProfileBodyActivityLevel)[keyof typeof UpdateProfileBodyActivityLevel];

export const UpdateProfileBodyActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
} as const;

export interface UpdateProfileBody {
  name?: string;
  weightKg?: number;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  ageYears?: number | null;
  goal?: UpdateProfileBodyGoal;
  activityLevel?: UpdateProfileBodyActivityLevel;
  dietaryRestrictions?: string[];
  preferredProteins?: string[];
}

export type MealMealType = (typeof MealMealType)[keyof typeof MealMealType];

export const MealMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
  pre_workout: "pre_workout",
  post_workout: "post_workout",
} as const;

export interface MealIngredient {
  name: string;
  quantityG: number;
  proteinG: number;
  carbsG: number;
  fatsG: number;
  caloriesKcal: number;
}

export interface Meal {
  name: string;
  mealType: MealMealType;
  ingredients: MealIngredient[];
  proteinG: number;
  carbsG: number;
  fatsG: number;
  caloriesKcal: number;
  /** @nullable */
  notes: string | null;
}

/**
 * @nullable
 */
export type DayPlanWorkoutType =
  | (typeof DayPlanWorkoutType)[keyof typeof DayPlanWorkoutType]
  | null;

export const DayPlanWorkoutType = {
  push: "push",
  pull: "pull",
  legs: "legs",
  rest: "rest",
} as const;

export interface DayPlan {
  dayNumber: number;
  dayName: string;
  /** @nullable */
  workoutType: DayPlanWorkoutType;
  meals: Meal[];
  totalProteinG: number;
  totalCarbsG: number;
  totalFatsG: number;
  totalCaloriesKcal: number;
}

export interface MealPlan {
  id: number;
  profileId: number;
  weekNumber: number;
  days: DayPlan[];
  targetProteinG: number;
  targetCaloriesKcal: number;
  createdAt: string;
}

export interface CreateMealPlanBody {
  profileId: number;
  weekNumber?: number;
}

export interface MealPlanSummary {
  mealPlanId: number;
  avgDailyProteinG: number;
  avgDailyCaloriesKcal: number;
  avgDailyCarbsG: number;
  avgDailyFatsG: number;
  proteinGoalMetDays: number;
  topProteinSources: string[];
  weeklyProteinG: number;
  weeklyCaloriesKcal: number;
}

export type IngredientCategory =
  (typeof IngredientCategory)[keyof typeof IngredientCategory];

export const IngredientCategory = {
  dairy: "dairy",
  legume: "legume",
  soy: "soy",
  supplement: "supplement",
  grain: "grain",
  vegetable: "vegetable",
  nut_seed: "nut_seed",
} as const;

export interface Ingredient {
  id: number;
  name: string;
  category: IngredientCategory;
  proteinPer100g: number;
  carbsPer100g: number;
  fatsPer100g: number;
  caloriesPer100g: number;
  isVegetarian: boolean;
  isVegan: boolean;
}

export type MacroTargetInputGoal =
  (typeof MacroTargetInputGoal)[keyof typeof MacroTargetInputGoal];

export const MacroTargetInputGoal = {
  muscle_gain: "muscle_gain",
  fat_loss: "fat_loss",
  maintenance: "maintenance",
} as const;

export type MacroTargetInputActivityLevel =
  (typeof MacroTargetInputActivityLevel)[keyof typeof MacroTargetInputActivityLevel];

export const MacroTargetInputActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
} as const;

export interface MacroTargetInput {
  weightKg: number;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  ageYears?: number | null;
  goal: MacroTargetInputGoal;
  activityLevel: MacroTargetInputActivityLevel;
}

export interface MacroTargets {
  dailyCaloriesKcal: number;
  dailyProteinG: number;
  dailyCarbsG: number;
  dailyFatsG: number;
  proteinPerKg: number;
}

export type ListMealPlansParams = {
  profileId?: number;
};