# lib/api-zod/src/generated/types/macroTargetInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * High-Protein Vegetarian Macro Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { MacroTargetInputActivityLevel } from "./macroTargetInputActivityLevel";
import type { MacroTargetInputGoal } from "./macroTargetInputGoal";

export interface MacroTargetInput {
  weightKg: number;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  ageYears?: number | null;
  goal: MacroTargetInputGoal;
  activityLevel: MacroTargetInputActivityLevel;
}