# lib/api-zod/src/generated/types/meal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * High-Protein Vegetarian Macro Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { MealIngredient } from "./mealIngredient";
import type { MealMealType } from "./mealMealType";

export interface Meal {
  name: string;
  mealType: MealMealType;
  ingredients: MealIngredient[];
  proteinG: number;
  carbsG: number;
  fatsG: number;
  caloriesKcal: number;
  /** @nullable */
  notes: string | null;
}