# lib/api-zod/src/generated/types/mealPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * High-Protein Vegetarian Macro Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { DayPlan } from "./dayPlan";

export interface MealPlan {
  id: number;
  profileId: number;
  weekNumber: number;
  days: DayPlan[];
  targetProteinG: number;
  targetCaloriesKcal: number;
  createdAt: Date;
}