# lib/api-zod/src/generated/types/mealPlanSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * High-Protein Vegetarian Macro Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface MealPlanSummary {
  mealPlanId: number;
  avgDailyProteinG: number;
  avgDailyCaloriesKcal: number;
  avgDailyCarbsG: number;
  avgDailyFatsG: number;
  proteinGoalMetDays: number;
  topProteinSources: string[];
  weeklyProteinG: number;
  weeklyCaloriesKcal: number;
}