# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * High-Protein Vegetarian Macro Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileActivityLevel } from "./profileActivityLevel";
import type { ProfileGoal } from "./profileGoal";

export interface Profile {
  id: number;
  name: string;
  weightKg: number;
  /** @nullable */
  heightCm: number | null;
  /** @nullable */
  ageYears: number | null;
  goal: ProfileGoal;
  activityLevel: ProfileActivityLevel;
  dietaryRestrictions: string[];
  preferredProteins: string[];
  createdAt: Date;
  updatedAt: Date;
}